Refactor unit tests to reduce duplicated tests for configurations. This reduces the redundancy and overhead in tests by leveraging a macro to run a given test accross multiple configurations.
