+++ src/tests.rs
@@ -1,0 +1,2 @@
+mod paging_tests;
+pub(crate) mod test_page_allocator;
